Fixed bug in airflowInit and airflowUpdate functions.
No need to create dag folder and copy there clean_dag_run.py
anymore as it's done by cwl-airflow init command. Although,
we still need to delete the original clean_dag_run.py otherwise
it won't be updated.
We also need to make sure the process_report connection is
updated. For that we remove and add it again. All command run
with runCommands should be executed sequentially, therefore
theay are combined with &&