Unified macOS and Ubuntu portable versions
Both macOS and Ubuntu portable scidap satellite installations use the same PM2 configuration generated by configure.js script based on the scidap_settings.json. The structure of scidap_settings.json is similar to what we had in the config.json (electron store), but was extended with additional keys. We use the latest Airflow 2.0.0 with PostrgeSQL backend. Check Notes for developers section in Readme for updates and known issues information